Fix eslint installation
Eslint had a conflict because it was installed both globally and locally. Deinsatlled it and reinstalled it globally, along with the necessary plugins: import, node, react. They need to be installed globally as well. Currently, they are installed both globally and locally, that is why they are listed in the devDependencies in package.json.