Add new recource functionality
The homepage now has "Add New" button. When clicked, it goes to the app.get('/add') route, which renders the add.html view. There is a form for adding a new resource, that gets submitted using the post method and it is routed to app.post('/add'). That route renders the list view and displays all the resources in the array. This functionality uses the existing getRecource and listRecource methods.