Set up React environment
Import react, webpack and babel dependencies. Set up webpack and babel. Webpack creates a bundle.js file that is hooked in a script tag on the index.html page in the view, and renders React from src/index.js as an entry point. Create a route (/react) to test the configuration.